update my meal

+++ app/controllers/my_meals_controller.rb
@@ -18,14 +18,16 @@ class MyMealsController < ApplicationController
 
     if @my_meal.save
       params[:ingredients].each do |i|
-        if !i[:ingredient][:foods].nil? || !i[:ingredient][:measures].nil? || !i[:ingredient][:qty].nil?
+        if i[:ingredient][:foods] != '' && i[:ingredient][:measures] != '' && i[:ingredient][:qty] != ''
           food = Food.find_by_name(i[:ingredient][:foods])
           measure = Measure.find_by_name(i[:ingredient][:measures])
+          description = i[:ingredient][:description]
           qty = i[:ingredient][:qty]
           ingredient = Ingredient.create(food_id: food.id,
                                         measure_id: measure.id,
                                         my_meal_id: @my_meal.id,
-                                        qty: qty)
+                                        qty: qty,
+                                        description: description)
           ingredient.save
         else
           # check how to handle the error
@@ -38,28 +40,38 @@ class MyMealsController < ApplicationController
   end
 
   def update
-    # my_meal = MyMeal.find(params[:id])
-    # my_meal.update_attributes(my_meal_params)
-    # if my_meal.avatar.attached?
-    #   my_meal.strMealThumb = url_for(my_meal.avatar)
-    # end
-    # if my_meal.save
-    #   redirect_to my_meals_path
-    # else
-    #   render :edit, locals: { notice: my_meal.errors.full_messages }
-    # end
+    my_meal = MyMeal.find(params[:id])
+    my_meal.update_attributes(my_meal_params)
+    if my_meal.avatar.attached?
+      my_meal.mealThumb = url_for(my_meal.avatar)
+    end
+    if my_meal.save
+      params[:ingredients].each do |i|
+        if i[:ingredient][:foods] != '' && i[:ingredient][:measures] != '' && i[:ingredient][:qty] != ''
+          food = Food.find_by_name(i[:ingredient][:foods])
+          measure = Measure.find_by_name(i[:ingredient][:measures])
+          description = i[:ingredient][:description]
+          qty = i[:ingredient][:qty]
+          ingredient = Ingredient.create(food_id: food.id,
+                                        measure_id: measure.id,
+                                        my_meal_id: my_meal.id,
+                                        qty: qty,
+                                        description: description)
+          ingredient.save
+        else
+          # check how to handle the error
+        end
+      end
+      redirect_to my_meals_path
+    else
+      render :edit, locals: { notice: my_meal.errors.full_messages }
+    end
   end
 
   def edit
-  #   @my_meal = MyMeal.find(params[:id])
-  # end
-
-  # def destroy
-  #   my_meal = MyMeal.find(params[:id])
-  #   my_meal.destroy
-  #   respond_to do |format|
-  #     format.json { render json: { notice: 'ok' } }
-  #   end
+    @my_meal = MyMeal.find(params[:id])
+    @foods = Food.all
+    @measures = Measure.all
   end
 
   def show
@@ -69,9 +81,13 @@ class MyMealsController < ApplicationController
 
   def destroy
     my_meal = MyMeal.find(params[:id])
-    my_meal.destroy
     respond_to do |format|
-      format.json { render json: { notice: 'ok' } }
+      if my_meal.weeks.length > 0
+        format.json { render json: { notice: "You can't delete this meal because its used in a week program" } }
+      else
+        my_meal.destroy
+        format.json { render json: { notice: 'ok' } }
+      end
     end
   end
